refactor: delegate schema manager creation responsibility to ClientAdapter
