+++ zeebe/exporters/camunda-exporter/src/main/java/io/camunda/exporter/schema/ElasticsearchEngineClient.java
@@ -329,9 +329,4 @@ public class ElasticsearchEngineClient implements SearchEngineClient {
           e);
     }
   }
-
-  public enum MappingSource {
-    INDEX,
-    INDEX_TEMPLATE
-  }
 }
